+++ include/Camera.h
@@ -11,6 +11,7 @@
 #include <glm/vec2.hpp>
 #include <glm/vec3.hpp>
 #include <glm/mat4x4.hpp>
+#include <glm/gtc/quaternion.hpp>
 
 #include "RoomData.h"
 
@@ -18,15 +19,17 @@ class Camera {
   public:
 
     static void reset();
+    static bool update();
 
     static void handleAction(ActionEvents action, bool isFinished);
     static void handleMouseMotion(int x, int y);
-    static bool update();
+    static void handleControllerAxis(float value, KeyboardButton axis);
+    static void handleControllerButton(KeyboardButton button, bool released);
 
     static void setPosition(glm::vec3 p) { pos = p; }
     static glm::vec3 getPosition() { return pos; }
 
-    static glm::vec2 getRotation() { return rot; }
+    static glm::vec2 getRotation();
     static glm::mat4 getProjectionMatrix() { return projection; }
     static glm::mat4 getViewMatrix() { return view; }
 
@@ -43,10 +46,12 @@ class Camera {
     static void displayFrustum(glm::mat4 MVP);
 
   private:
+    static void calculateFrustumPlanes();
+
     static glm::vec3 pos;
-    static glm::vec2 rot;
-    static glm::vec3 lastPos;
-    static glm::vec2 lastRot;
+    static glm::quat quaternion;
+    static glm::vec3 posSpeed;
+    static glm::vec2 rotSpeed;
     static glm::vec2 lastSize;
     static glm::mat4 projection;
     static glm::mat4 view;