+++ src/Camera.cpp
@@ -6,9 +6,13 @@
  * \author xythobuz
  */
 
+#include <limits>
+#include <glm/gtc/epsilon.hpp>
 #include <glm/gtc/matrix_transform.hpp>
+#include <glm/gtx/quaternion.hpp>
 
 #include "global.h"
+#include "RunTime.h"
 #include "system/Window.h"
 #include "Camera.h"
 
@@ -28,9 +32,134 @@
 #define FBR 6
 #define FTR 7
 
-class Plane {
+const static float fov = 45.0f;
+const static float nearDist = 0.1f;
+const static float farDist = 75000.0f;
+const static float maxSpeed = 2048.0f;
+const static float controllerViewFactor = 384.0f;
+
+const static glm::vec3 rightUnit(1.0f, 0.0f, 0.0f);
+const static glm::vec3 upUnit(0.0f, 1.0f, 0.0f);
+const static glm::vec3 dirUnit(0.0f, 0.0f, -1.0f);
+
+glm::vec3 Camera::pos(0.0f, 0.0f, 0.0f);
+glm::quat Camera::quaternion(glm::vec3(0.0f, 0.0f, 0.0f));
+glm::vec3 Camera::posSpeed(0.0f, 0.0f, 0.0f);
+glm::vec2 Camera::rotSpeed(0.0f, 0.0f);
+glm::vec2 Camera::lastSize(0.0f, 0.0f);
+glm::mat4 Camera::projection(1.0f);
+glm::mat4 Camera::view(1.0f);
+float Camera::rotationDeltaX = 0.75f;
+float Camera::rotationDeltaY = 0.75f;
+bool Camera::updateViewFrustum = true;
+
+void Camera::reset() {
+    pos = glm::vec3(0.0f, 0.0f, 0.0f);
+    quaternion = glm::quat(glm::vec3(0.0f, 0.0f, 0.0f));
+    posSpeed = glm::vec3(0.0f, 0.0f, 0.0f);
+    rotSpeed = glm::vec2(0.0f, 0.0f);
+    lastSize = glm::vec2(0.0f, 0.0f);
+    projection = glm::mat4(1.0f);
+    view = glm::mat4(1.0f);
+}
+
+void Camera::handleAction(ActionEvents action, bool isFinished) {
+    float factor = 1.0f;
+    if (isFinished)
+        factor = -1.0f;
+
+    if (action == forwardAction) {
+        posSpeed += dirUnit * maxSpeed * factor;
+    } else if (action == backwardAction) {
+        posSpeed -= dirUnit * maxSpeed * factor;
+    } else if (action == leftAction) {
+        posSpeed += rightUnit * maxSpeed * factor;
+    } else if (action == rightAction) {
+        posSpeed -= rightUnit * maxSpeed * factor;
+    } else if (action == jumpAction) {
+        posSpeed += upUnit * maxSpeed * factor;
+    } else if (action == crouchAction) {
+        posSpeed -= upUnit * maxSpeed * factor;
+    }
+}
+
+void Camera::handleMouseMotion(int x, int y) {
+    if (x != 0)
+        quaternion = glm::quat(upUnit * rotationDeltaX * float(x)) * quaternion;
+
+    if (y != 0)
+        quaternion = glm::quat(quaternion * -rightUnit * rotationDeltaY * float(y)) * quaternion;
+}
+
+void Camera::handleControllerAxis(float value, KeyboardButton axis) {
+    if (glm::epsilonEqual(value, 0.0f, 0.01f))
+        value = 0.0f;
+
+    if (axis == leftXAxis) {
+        posSpeed.x = -maxSpeed * value;
+    } else if (axis == leftYAxis) {
+        posSpeed.z = maxSpeed * value;
+    } else if (axis == rightXAxis) {
+        rotSpeed.x = controllerViewFactor * value;
+    } else if (axis == rightYAxis) {
+        rotSpeed.y = controllerViewFactor * value;
+    }
+}
+
+void Camera::handleControllerButton(KeyboardButton button, bool released) {
+    if (button == aButton) {
+        handleAction(jumpAction, released);
+    } else if (button == bButton) {
+        handleAction(crouchAction, released);
+    } else if (button == padUp) {
+        handleAction(forwardAction, released);
+    } else if (button == padDown) {
+        handleAction(backwardAction, released);
+    } else if (button == padLeft) {
+        handleAction(leftAction, released);
+    } else if (button == padRight) {
+        handleAction(rightAction, released);
+    }
+}
+
+bool Camera::update() {
+    glm::vec2 size(getWindow().getWidth(), getWindow().getHeight());
+
+    if (lastSize != size) {
+        //! \fixme TODO instead of mirroring the Y axis in the shader, scale with -1 here
+        projection = glm::perspective(fov, size.x / size.y, nearDist, farDist);
+        lastSize = size;
+    }
+
+    float dT = getRunTime().getLastFrameTime();
+    pos += quaternion * posSpeed * dT;
+
+    if (glm::epsilonNotEqual(rotSpeed.x, 0.0f, 0.01f))
+        quaternion = glm::quat(upUnit * rotationDeltaX * rotSpeed.x * dT) * quaternion;
+
+    if (glm::epsilonNotEqual(rotSpeed.y, 0.0f, 0.01f))
+        quaternion = glm::quat(quaternion * -rightUnit * rotationDeltaY * rotSpeed.y * dT) * quaternion;
+
+    glm::mat4 translate = glm::translate(glm::mat4(1.0f), pos);
+    glm::mat4 rotate = glm::toMat4(quaternion);
+    view = glm::inverse(translate * rotate);
+
+    if (updateViewFrustum)
+        calculateFrustumPlanes();
+
+    return updateViewFrustum;
+}
+
+glm::vec2 Camera::getRotation() {
+    glm::vec3 euler = glm::eulerAngles(quaternion);
+    return glm::vec2(euler.y, euler.x);
+}
+
+// ----------------------------------------------------------------------------
+
+class FrustumPlane {
   public:
-    Plane() : normal(glm::vec3(0.0f, 0.0f, 0.0f)), d(0.0f) { }
+    FrustumPlane() : normal(glm::vec3(0.0f, 0.0f, 0.0f)), d(0.0f) { }
     void set(glm::vec3 v1, glm::vec3 v2, glm::vec3 v3) {
         normal = glm::normalize(glm::cross(v3 - v2, v1 - v2));
         d = -glm::dot(normal, v2);
@@ -45,23 +174,7 @@ class Plane {
 
 // ----------------------------------------------------------------------------
 
-const static float fov = 45.0f;
-const static float nearDist = 0.1f;
-const static float farDist = 75000.0f;
-const static float freeCameraStep = 256.0f;
-
-glm::vec3 Camera::pos(0.0f, 0.0f, 0.0f);
-glm::vec2 Camera::rot(0.0f, 0.0f);
-glm::vec3 Camera::lastPos(1.0f, 0.0f, 0.0f);
-glm::vec2 Camera::lastRot(1.0f, 0.0f);
-glm::vec2 Camera::lastSize(0.0f, 0.0f);
-glm::mat4 Camera::projection(1.0f);
-glm::mat4 Camera::view(1.0f);
-float Camera::rotationDeltaX = 0.75f;
-float Camera::rotationDeltaY = 0.75f;
-bool Camera::updateViewFrustum = true;
-static Plane planes[6];
-
+static FrustumPlane planes[6];
 static glm::vec3 frustumColors[6] = {
     glm::vec3(1.0f, 0.0f, 0.0f), // NEAR, red
     glm::vec3(0.0f, 1.0f, 0.0f), // FAR, green
@@ -72,109 +185,7 @@ static glm::vec3 frustumColors[6] = {
 };
 static glm::vec3 frustumVertices[8];
 
-void Camera::reset() {
-    pos = glm::vec3(0.0f, 0.0f, 0.0f);
-    rot = glm::vec2(0.0f, 0.0f);
-    lastPos = glm::vec3(1.0f, 0.0f, 0.0f);
-    lastRot = glm::vec2(1.0f, 0.0f);
-    lastSize = glm::vec2(0.0f, 0.0f);
-    projection = glm::mat4(1.0f);
-    view = glm::mat4(1.0f);
-}
-
-void Camera::handleAction(ActionEvents action, bool isFinished) {
-    if (isFinished)
-        return;
-
-    glm::vec3 dir(
-        glm::cos(rot.y) * glm::sin(rot.x),
-        glm::sin(rot.y),
-        glm::cos(rot.y) * glm::cos(rot.x)
-    );
-    glm::vec3 right(
-        glm::sin(rot.x - glm::pi<float>() / 2.0f),
-        0.0f,
-        glm::cos(rot.x - glm::pi<float>() / 2.0f)
-    );
-    glm::vec3 up = glm::vec3(0.0f, 1.0f, 0.0f);
-
-    if (action == forwardAction) {
-        pos += dir * freeCameraStep;
-    } else if (action == backwardAction) {
-        pos -= dir * freeCameraStep;
-    } else if (action == leftAction) {
-        pos += right * freeCameraStep;
-    } else if (action == rightAction) {
-        pos -= right * freeCameraStep;
-    } else if (action == jumpAction) {
-        pos += up * freeCameraStep;
-    } else if (action == crouchAction) {
-        pos -= up * freeCameraStep;
-    } else if (action == useAction) {
-    } else if (action == holsterAction) {
-    } else if (action == walkAction) {
-    }
-}
-
-void Camera::handleMouseMotion(int x, int y) {
-    while (x > 0) {
-        rot.x += rotationDeltaX;
-        x--;
-    }
-    while (x < 0) {
-        rot.x -= rotationDeltaX;
-        x++;
-    }
-    while (y > 0) {
-        if (rot.y > -(glm::pi<float>() / 2.0f)) {
-            rot.y -= rotationDeltaY;
-        }
-        y--;
-    }
-    while (y < 0) {
-        if (rot.y < (glm::pi<float>() / 2.0f)) {
-            rot.y += rotationDeltaY;
-        }
-        y++;
-    }
-
-    while (rot.x > (glm::pi<float>() * 2.0f))
-        rot.x -= glm::pi<float>() * 2.0f;
-
-    while (rot.x < -(glm::pi<float>() * 2.0f))
-        rot.x += glm::pi<float>() * 2.0f;
-}
-
-bool Camera::update() {
-    glm::vec2 size(getWindow().getWidth(), getWindow().getHeight());
-
-    if ((lastPos == pos) && (lastRot == rot) && (lastSize == size))
-        return false;
-
-    if (lastSize != size) {
-        //! \fixme TODO instead of mirroring the axes in the shader, scale here
-        projection = glm::perspective(fov, size.x / size.y, nearDist, farDist);
-        lastSize = size;
-    }
-
-    glm::vec3 dir(
-        glm::cos(rot.y) * glm::sin(rot.x),
-        glm::sin(rot.y),
-        glm::cos(rot.y) * glm::cos(rot.x)
-    );
-
-    glm::vec3 right(
-        glm::sin(rot.x - glm::pi<float>() / 2.0f),
-        0.0f,
-        glm::cos(rot.x - glm::pi<float>() / 2.0f)
-    );
-
-    glm::vec3 up = glm::cross(right, dir);
-    view = glm::lookAt(pos, pos + dir, up);
-
-    if (!updateViewFrustum)
-        return false;
-
+void Camera::calculateFrustumPlanes() {
     glm::mat4 combo = projection * view;
 
     // Calculate frustum corners to display them
@@ -200,10 +211,6 @@ bool Camera::update() {
     planes[RIGHT].set(frustumVertices[NBR], frustumVertices[NTR], frustumVertices[FBR]);
     planes[NEAR].set(frustumVertices[NTL], frustumVertices[NTR], frustumVertices[NBR]);
     planes[FAR].set(frustumVertices[FTR], frustumVertices[FTL], frustumVertices[FBL]);
-
-    lastPos = pos;
-    lastRot = rot;
-    return true;
 }
 
 bool Camera::boxInFrustum(BoundingBox b) {